skl_backend/requirements.txt
Django
djangorestframework
gunicorn
Pillow
django-ckeditor
tzdata
django-colorfield